Bug 129 COM Crash.
There were two issues -
ObjFromVariant was setting the parameter VARIANT type even in case of errors.
In the case an IDispatch param was wrong type, this resulted in the VARIANT
being freed (crash) as VT_DISPATCH even though it was not. Versions prior
to 4.0b53 behaved did not suffer the crash as they did not decr ref VT_DISPATCH
parameters.

The parameter itself was the wrong type because the comobj was not converted
to a IDispatch type in the Automation proxy. This was because loop variable
for parameter number was never incremented resulting in parameter type always
being retrieved from the first parameter.